Move built-ins parsing and ingestion to `slang_solidity` crate
Since the built-ins file needs to be pre-processed to transform the symbols as
to ensure no conflicts can occur with user code, adding the built-ins requires a
couple of manual steps that were replicated in every construction of Solidity
bindings API. By encapsulating this functionality in the `slang_solidity` crate
we remove a source of user error and make it easier to make changes to the
built-ins ingestion code.
