+++ crates/solidity/testing/perf/src/tests/definitions.rs
@@ -8,7 +8,7 @@ pub struct Dependencies {
 }
 
 pub fn setup() -> Dependencies {
-    let bindings = super::init_bindings::run(super::init_bindings::setup());
+    let bindings = super::init_bindings::run();
     let files = super::parser::run(super::parser::setup());
 
     Dependencies { bindings, files }
